[Fix] 타입 오류 수정 및 위치 이동
