+++ darwin/src/api/product/deleteProduct.ts
@@ -1,7 +1,7 @@
 import {useMutation, useQueryClient} from "@tanstack/react-query";
 import apiClient from "@/api/apiClient";
 
-const deleteProduct = async (id: number) => {
+const deleteProduct = async (id: string) => {
     const response = await apiClient.delete(`/api/v1/products/${id}`);
     return response.data;
 };
@@ -9,7 +9,7 @@ const deleteProduct = async (id: number) => {
 export const useDeleteProduct = () => {
     const queryClient = useQueryClient();
     return useMutation({
-        mutationFn: (id: number) => deleteProduct(id),
+        mutationFn: (id: string) => deleteProduct(id),
         onSuccess: () => queryClient.invalidateQueries({ queryKey: ["products"] }),
     });
 };
