+++ darwin/src/api/product/getProduct.ts
@@ -1,14 +1,23 @@
 import { useQuery } from "@tanstack/react-query";
 
 import apiClient from "@/api/apiClient";
-import { ProductDetails } from "./ProductApiTypes";
 
-const getProduct = async (productId: number) => {
+export interface ProductDetails {
+    id: string;
+    name: string;
+    description: string;
+    stock: number;
+    createdAt: string;
+    thumbnailUrl: string;
+    detailFileUrls: string[];
+}
+
+const getProduct = async (productId: string) => {
     const response = await apiClient.get(`/api/v1/products/${productId}`);
     return response.data;
 };
 
-export const useGetProduct = (productId: number) => {
+export const useGetProduct = (productId: string) => {
     return useQuery<ProductDetails>({
         queryKey: ["productId", productId],
         queryFn: () => {
