+++ darwin/src/api/product/getProducts.ts
@@ -1,7 +1,13 @@
 import { useQuery } from "@tanstack/react-query";
 
-import { Product } from "./ProductApiTypes";
 import apiClient from "@/api/apiClient";
+
+export interface Product {
+    id: string;
+    productName: string;
+    thumbnailUrl: string;
+    stock: number;
+}
 
 const getProducts = async () => {
     const response = await apiClient.get(`/api/v1/products`);
